removed back buttons on mobile, returned title on mobile, refactor header
