+++ frontend/assets/less/variables.less
@@ -59,7 +59,7 @@
   //BODY PADDINGS
   --padding-side: 12px;
   --padding-vertical: 16px;
-  --header-height: 52px;
+  --header-height:64px;
 
   @media (min-width: 768px) {
     --padding-side: 40px;
