+++ frontend/components/header/Tablet.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts"></script>
+
+<template>
+	<HeaderUiWrapper>
+		<template #left>
+			<HeaderUiLogo />
+		</template>
+
+		<template #right>
+			<HeaderUiCreateButton />
+			<HeaderLanguageSelector />
+			<HeaderNavigationMobile />
+		</template>
+	</HeaderUiWrapper>
+</template>
