+++ frontend/layouts/default.vue
@@ -3,6 +3,7 @@ import { RouteNameEnum } from '../constants/enums/route';
 import { getRouteName } from '../utils';
 const route = useRoute();
 const mobile = inject('mobile');
+const tablet = inject('tablet');
 
 const isMobileFooterVisible = computed(() => {
 	const currentRouteName = getRouteName(route.name);
@@ -19,7 +20,9 @@ const isMobileFooterVisible = computed(() => {
 
 <template>
 	<div class="root">
-		<HeaderCommon :has-back-button="getRouteName(route.name) === RouteNameEnum.EVENT" />
+		<HeaderMobile v-if="mobile" />
+		<HeaderTablet v-if="tablet" />
+		<HeaderDesktop v-if="!mobile && !tablet" />
 		<slot> </slot>
 
 		<FooterCommon v-if="!mobile || (mobile && isMobileFooterVisible)" />
